pyproject.toml
[project]
name = "nonebot-plugin-argot"
version = "0.1.5"
description = "NoneBot 暗语"
authors = [
    {name = "KomoriDev", email = "mute231010@gmail.com"},
]
dependencies = [
    "nonebot2>=2.3.2",
    "nonebot-plugin-alconna>=0.51.1",
    "nonebot-plugin-orm>=0.7.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "aiofiles>=24.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
adapters = [
    "nonebot-adapter-discord>=0.1.8",
    "nonebot-adapter-dodo>=0.2.1",
    "nonebot-adapter-feishu>=2.6.2",
    "nonebot-adapter-kaiheila>=0.3.4",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-adapter-qq>=1.5.0",
    "nonebot-adapter-red>=0.9.0",
    "nonebot-adapter-satori>=0.12.3",
    "nonebot-adapter-telegram>=0.1.0b17",
]
tests = [
    "nonebug>=0.3.7",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.23.8",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.1",
    "black>=24.4.2",
    "isort>=5.13.2",
    "pre-commit>=3.7.1",
    "nonebot2[fastapi]>=2.3.2",
    "nonebot-plugin-orm[default]>=0.7.6",
]

[tool.nonebot]
plugins = ["nonebot_plugin_argot"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
